address more review comments
